bugfix: orphaned binary gets killed on GET /focus
After the original neovim process is closed, doing a GET /focus?a=b
causes the binary to die. This issue doesn't occur if
	g:nvim_ghost_logging_enabled = 1
which leads me to think that it might be caused by stdout/stderr error.

For eg. when an application wants to write something to a file
descriptor that has been closed already, it causes an error.

But what might cause the stdout/stderr to get closed on closing of nvim?
Hmm... remember the stdout/stderr callbacks we set while spawning the
function, the callbacks that are meant to log the output? Well, when
nvim is closed, nobody is there to handle the callbacks anymore! So,
maybe that's causing the binary to get killed?

We checked by setting s:joblog_arguments = {}, and sure enough, the bug
is gone! Hooray!

So, now we set s:joblog_arguments to {} if g:nvim_ghost_logging_enabled
is false.