[BUGFIX] nvim instance can't connect after install
If there's only one neovim instance, and the binary is out-of-date, the
neovim instance installs the binary, but doesn't start it because it
receives no UIEnter event. So, the binary ends up never starting at all,
causing GhostText to be unable to connect until a new neovim instance is
started.